Add GetCurrentDIDMethod component
